requirements.txt
pytest-playwright==0.6.2
requests==2.32.3
python-dotenv==1.0.1
